fix: Changed structure of the AuthorizationPage
