--- postcss.config.mjs
@@ -1,5 +1,0 @@
-const config = {
-    plugins: ["@tailwindcss/postcss"],
-};
-
-export default config;
