+++ src/app/layout.tsx
@@ -1,8 +1,5 @@
-import ProtectedRoute from "@components/ProtectedRoute";
 import StoreWrapper from "@components/StoreWrapper";
-import store from "@store/index";
 import { type Metadata } from "next";
-import { Provider } from "react-redux";
 
 export const metadata: Metadata = {
     title: "Market frontend template",
